backend/requirements.txt
flask
flask-cors
pymongo
PyPDF2
scikit-learn
pandas
python-dotenv
werkzeug
